autoTest/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
python-jose
passlib
python-multipart
email-validator 